test/mnist-cnn-demo/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
numpy==1.26.4
tensorflow==2.15.1
tf2onnx==1.16.1
opentelemetry-sdk==1.25.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-logging==0.46b0
prometheus-client==0.20.0
